% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{startShinyApp}
\alias{startShinyApp}
\title{Runs the Shiny app.}
\usage{
startShinyApp(port = 19901, is_local_development = TRUE)
}
\value{
This starts up the shiny app interface
}
\description{
Runs the Shiny app.
}
